requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.10.2
    # via -r requirements.in
anyio==3.6.2
    # via starlette
asyncpg==0.27.0
    # via -r requirements.in
click==8.1.3
    # via uvicorn
decorator==5.1.1
    # via validators
fastapi==0.95.0
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
importlib-metadata==6.6.0
    # via alembic
importlib-resources==5.12.0
    # via alembic
mako==1.2.4
    # via alembic
markupsafe==2.1.2
    # via mako
psycopg2-binary==2.9.6
    # via -r requirements.in
pydantic==1.10.7
    # via fastapi
python-dotenv==1.0.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.8
    # via
    #   -r requirements.in
    #   alembic
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via
    #   alembic
    #   pydantic
    #   sqlalchemy
    #   starlette
uvicorn==0.21.1
    # via -r requirements.in
validators==0.20.0
    # via -r requirements.in
zipp==3.15.0
    # via
    #   importlib-metadata
    #   importlib-resources
